confirm logout

+++ jobs/session/bin/logout.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-SLEEP_PID=$(pgrep -s 0 sleep)
-kill -9 ${SLEEP_PID}
+if zenity --question --text="<span size=\"x-large\">Are you sure you want to log out?</span>" --title="Log out"; then
+  SLEEP_PID=$(pgrep -s 0 sleep)
+  kill -9 ${SLEEP_PID}
+fi
